Move Address back to components dir, add lifecycle check to initialize editableAddress
